+++ CHANGES.md
@@ -13,10 +13,13 @@ NEXT MILESTONE
 - Add API for `cumulative` when only starts are variable
 - Add decomposition of cumulative: `model.cumulativeTimeDecomp(...)`
 - Logical expression XOR manages more than 2 variables
+- Add new API to IOutputFactory (to Gephi and to Graphviz)
+- Add constraint network output (to gexf format), see `solver.constraintNetworkToGephi(file)`
 
 ### Deprecated API (to be removed in next release):
 
 ### Closed issues and pull requests:
+\#600, #601, #602
 
 
 4.0.8 - 23 Jul 2018